[Keychron C1] Cleanup keymaps
